Move Open* under baaahs.show.live.
