+++ src/commonMain/kotlin/baaahs/show/PatchHolder.kt
@@ -1,0 +1,8 @@
+package baaahs.show
+
+interface PatchHolder {
+    val title: String
+    val patches: List<Patch>
+    val eventBindings: List<EventBinding>
+    val controlLayout: Map<String, List<ControlRef>>
+}
