+++ src/jsMain/kotlin/baaahs/app/ui/ControlsPalette.kt
@@ -1,10 +1,10 @@
 package baaahs.app.ui
 
-import baaahs.OpenShow
 import baaahs.app.ui.controls.ControlDisplay
 import baaahs.app.ui.controls.SpecialControlProps
 import baaahs.app.ui.controls.control
 import baaahs.show.ShowBuilder
+import baaahs.show.live.OpenShow
 import baaahs.ui.*
 import external.Direction
 import external.draggable
