+++ src/jsMain/kotlin/baaahs/app/ui/controls/SceneList.kt
@@ -1,12 +1,12 @@
 package baaahs.app.ui.controls
 
-import baaahs.OpenShow
 import baaahs.ShowState
 import baaahs.app.ui.Draggable
 import baaahs.app.ui.DropTarget
 import baaahs.app.ui.appContext
 import baaahs.show.PatchHolderEditor
 import baaahs.show.Show
+import baaahs.show.live.OpenShow
 import baaahs.ui.*
 import external.Direction
 import external.copyFrom
